rename speakers.js. add image. modify session
