+++ contents/speakers/speakers.js
@@ -1,10 +1,10 @@
 const speakers = [
-  "tanakataro1",
-  "tanakataro2"
+  "onoueyosuke",
+  "taguchiwataru"
 ];
 
 const speakersList = speakers.map(speaker => {
-  return require(`json-loader!yaml-loader!~/contents/speakers/${speaker}.yml`);
+  return require(`json-loader!yaml-loader!./${speaker}.yml`);
 });
 
 export const contents = speakersList;
