--- contents/speakers/tanakataro2.yml
@@ -1,10 +1,0 @@
-name: 田中 太郎
-company: 株式会社chatbox
-detail: 【140文字】株式会社 chatbox は大阪の小さなWeb制作会社です。 様々 なWeb制作の現場で、「Webのこれからを語り合う」をテーマにWebの 可能性を追求し続けています。 Web制作という「作る」業務だけでな く、Webの技術を「伝える」「語り合う」会社として、 制作現場。
-image: sapmle.png
-twitter: https://twitter.com/
-facebook: https://www.facebook.com/
-link: https://www.google.com/
-session:
-  title: 【40文字】作るだけなく、伝え語り合う会社でありたい作るだけなく、伝え語り合う会
-  detail: 【270文字】株式会社 chatbox は大阪の小さなWeb制作会社です。 様々なWeb制作の現場で、 「Webのこれからを語り合う」をテーマにWebの可能性を追求し続けています。 Web制作という 「作る」業務だけでなく、Webの技術を「伝える」「語り合う」会社として、 制作現場での技術 指導や、イベント・セミナーなどの交流活動にも積極的に取り組んでいます。 身の回りにありふ れたWebだからこそ、そこにしか無い、そこだけのWebの形を。 現場と、お客様と、ユーザー と、一緒に考えていくのがchatboxのWeb制作のあり方です。ああああああ
